requirements.txt
# --- Backend & API Framework ---
fastapi==0.136.0
uvicorn==0.45.0
pydantic==2.13.3
pydantic-settings==2.14.0
python-multipart==0.0.26

# --- Database & Migrations ---
sqlalchemy==2.0.49
# Note: psycopg2-binary wasn't in your list, but is needed for PostgreSQL
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Task Queue & Caching ---
celery==5.6.3
redis==5.0.4

# --- AI & LangChain Ecosystem ---
langchain==1.2.15
langchain-core==1.3.0
langchain-community==0.4.1
langchain-openai==1.1.16
langgraph==1.1.9
langsmith==0.7.33

# --- Vector Database & Embeddings ---
chromadb==1.5.8
tiktoken==0.12.0

# --- Document Processing & RAG ---
# This version is compatible with your Python 3.12 environment
unstructured[pdf]==0.22.22
ragas==0.4.3
datasets==4.8.4

# --- Integration & Tools ---
mcp==1.27.0
httpx==0.28.1
python-dotenv==1.2.2
python-telegram-bot==22.7